refactor: add readme and refactor some code in qrlink test suite
