+++ packages/vc-test-suite/tests/QrlinkEncrypted/qrlink-encrypted.test.ts
@@ -1,5 +1,4 @@
 import chai from 'chai';
-import { before } from 'node:test';
 import { decryptString } from '@govtechsg/oa-encryption';
 import { isURLEncoded, parseQRLink } from './helper';
 import { request } from '../../httpService';
@@ -9,29 +8,11 @@ import * as config from '../../config';
 const expect = chai.expect;
 
 describe('QR Link Verification', function () {
-  let qrLink = '';
-  let parsedLink = {
-    verify_app_address: '',
-    q: {
-      payload: {
-        uri: '',
-        key: '',
-        hash: '',
-      },
-    },
-  };
+  const qrLink = config.default.testSuites.QrLinkEncrypted.url;
+  const method = config.default.testSuites.QrLinkEncrypted.method;
+  const parsedLink = parseQRLink(qrLink);
 
   setupMatrix.call(this, [config.default.implementationName], 'Implementer');
-
-  before(() => {
-    try {
-      qrLink = require('./input/qr-link.json').url;
-      parsedLink = parseQRLink(qrLink);
-    } catch (error) {
-      // eslint-disable-next-line no-console
-      console.error(`QR Link: failed to read "qr-link.json".`);
-    }
-  });
 
   reportRow('QR link MUST be URL encoded', config.default.implementationName, function () {
     const data = isURLEncoded(qrLink);
@@ -53,8 +34,9 @@ describe('QR Link Verification', function () {
   reportRow('URI MUST be resolvable', config.default.implementationName, async function () {
     const res = await request({
       url: parsedLink.q.payload.uri,
-      method: 'GET',
+      method,
     });
+
     res.should.be.an('object');
   });
 
@@ -63,6 +45,7 @@ describe('QR Link Verification', function () {
   });
 
   reportRow('Hash MUST match the credential hash', config.default.implementationName, async function () {
+    // TODO: Implement this test case with hash comparison
     // const res = await request({
     //   url: parsedLink.q.payload.uri,
     //   method: 'GET',
@@ -84,7 +67,7 @@ describe('QR Link Verification', function () {
   reportRow('Key MUST decrypt the encrypted credential', config.default.implementationName, async function () {
     const res = await request({
       url: parsedLink.q.payload.uri,
-      method: 'GET',
+      method,
     });
     const stringifyVC = decryptString({
       ...res.document,
